move build email context
